func_80045C98 -4 -> -1 on one declaration keyword
**`s16 b`, not `s32 b` with an `(s16)` cast at the call.**  The cast folds
-- after `andi 0xFFFF` and `srl 8` gcc can prove the value is 0..255 -- and
the narrow declaration makes the truncation part of the assignment instead,
which brings back three of the four missing instructions: the `j` to the
join, the `else` arm's `addu`, and one of the two shifts.

That is the narrow-local family read for a new effect.  The standing rules
say a narrow declaration CREATES an instruction (func_8005AE68) or rotates
registers by creating a pseudo (func_80027060); here it creates a whole
control-flow shape, because the truncation gcc cannot fold keeps the
if/else from collapsing into an unconditional store plus a conditional
overwrite.

Measured and worse: `u16 b` with the cast (still -1), an `s16`-parameter
prototype for func_80049F10 (-4), and `(s16)(b & 0xFFFF)` at the call (-4).

The last -1 is the extension itself: retail does `sll`/`sra` at the join
and we copy the already-narrow `b` into `$a0`.  Retail's `b` is an SImode
value extended AT the call and ours is a 16-bit value extended at the
assignment, and no spelling found so far gives the first without the fold.

Also measured, no change: three more abs spellings on func_80058624's one
difference (`if`/`else`, ternary, and testing the copy rather than the
original) -- gcc propagates the original into the negation every time,
where retail negates the copy.

+++ parked/func_80045C98.c
@@ -3,7 +3,7 @@
 void func_80045C98(void) {
     u8 f;
     s32 a;
-    s32 b;
+    s16 b;
     s32 h;
 
     f = 0;
@@ -56,7 +56,7 @@ void func_80045C98(void) {
         } else {
             b = 0;
         }
-        func_80049F10((s16)b, (s16)b);
+        func_80049F10(b, b);
     }
 
     if (*(s16 *)(D_8009B45C + 0x512) != 0) {
